pyproject.toml
[project]
name = "snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-fasthtml>=0.12.1"]

[tool.pyright]
venvPath = "."
venv = ".venv"
